phase6_orchestrator/pyproject.toml
[tool.poetry]
name = "phase6-orchestrator"
version = "0.1.0"
description = "Phase 6: Single-file orchestrator for audiobook pipeline phases 1-5"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "phase6_orchestrator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0"
rich = "^13.0"
pyyaml = "^6.0"
charset-normalizer = "^3.4.3"
python-docx = "^1.2.0"
ebooklib = "^0.19"
beautifulsoup4 = "^4.14.2"
lxml = "^6.0.2"
readability-lxml = "^0.8.4.1"
pdf2image = "^1.17.0"
pypdf = "^6.1.1"
python-magic-bin = "^0.4.14"

[tool.poetry.scripts]
phase6 = "phase6_orchestrator.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
